backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.3.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
aiofiles>=0.7.0
jinja2>=3.0.1
websockets>=10.0
supabase>=2.0.0
alembic>=1.7.0 # Add alembic for database migrations
psycopg2-binary>=2.9.0 # Add postgres driver needed by alembic/sqlalchemydeepgram-sdk
requests==2.32.3
